feat: 添加全局主题上下文
